Add volume calculation of DifShells and DifBuffers
Volume needs to be calculated before a diffusion message with calcium concentration or buffer concentration is obtained. Volume is calculated at reinit and sometimes a diffusion message is received before the reinit resulting in a division by volume of zero. Due to that initial concentration in some difshells and difbuffers is not correct.